Move user routes over.
