+++ politeiawww/legacy/user_test.go
