+++ politeiawww/user.go
@@ -5,1049 +5,18 @@
 package main
 
 import (
-	"bytes"
-	"context"
-	"encoding/hex"
-	"errors"
 	"fmt"
 	"os"
-	"sort"
-	"strconv"
-	"strings"
-	"sync"
 	"time"
 
-	"github.com/decred/politeia/politeiad/api/v1/identity"
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
 	"github.com/decred/politeia/politeiawww/user"
-	"github.com/decred/politeia/util"
-	"github.com/google/uuid"
-	"golang.org/x/crypto/bcrypt"
 )
-
-// loginReply is used to pass the results of the login command between go
-// routines.
-type loginResult struct {
-	reply *www.LoginReply
-	err   error
-}
-
-// resetPassword is used to pass the results of the reset password command
-// between go routines.
-type resetPasswordResult struct {
-	reply www.ResetPasswordReply
-	err   error
-}
-
-// convertWWWUserFromDatabaseUser converts a user User to a www User.
-func convertWWWUserFromDatabaseUser(user *user.User) www.User {
-	return www.User{
-		ID:                              user.ID.String(),
-		Admin:                           user.Admin,
-		Email:                           user.Email,
-		Username:                        user.Username,
-		NewUserPaywallAddress:           user.NewUserPaywallAddress,
-		NewUserPaywallAmount:            user.NewUserPaywallAmount,
-		NewUserPaywallTx:                user.NewUserPaywallTx,
-		NewUserPaywallTxNotBefore:       user.NewUserPaywallTxNotBefore,
-		NewUserPaywallPollExpiry:        user.NewUserPaywallPollExpiry,
-		NewUserVerificationToken:        user.NewUserVerificationToken,
-		NewUserVerificationExpiry:       user.NewUserVerificationExpiry,
-		UpdateKeyVerificationToken:      user.UpdateKeyVerificationToken,
-		UpdateKeyVerificationExpiry:     user.UpdateKeyVerificationExpiry,
-		ResetPasswordVerificationToken:  user.ResetPasswordVerificationToken,
-		ResetPasswordVerificationExpiry: user.ResetPasswordVerificationExpiry,
-		LastLoginTime:                   user.LastLoginTime,
-		FailedLoginAttempts:             user.FailedLoginAttempts,
-		Deactivated:                     user.Deactivated,
-		Locked:                          userIsLocked(user.FailedLoginAttempts),
-		Identities:                      convertWWWIdentitiesFromDatabaseIdentities(user.Identities),
-		ProposalCredits:                 uint64(len(user.UnspentProposalCredits)),
-		EmailNotifications:              user.EmailNotifications,
-	}
-}
-
-// convertWWWIdentitiesFromDatabaseIdentities converts a user Identity to a www
-// Identity.
-func convertWWWIdentitiesFromDatabaseIdentities(identities []user.Identity) []www.UserIdentity {
-	userIdentities := make([]www.UserIdentity, 0, len(identities))
-	for _, v := range identities {
-		userIdentities = append(userIdentities,
-			convertWWWIdentityFromDatabaseIdentity(v))
-	}
-	return userIdentities
-}
-
-// convertWWWIdentityFromDatabaseIdentity converts a user Identity to a www
-// Identity.
-func convertWWWIdentityFromDatabaseIdentity(id user.Identity) www.UserIdentity {
-	return www.UserIdentity{
-		Pubkey: id.String(),
-		Active: id.IsActive(),
-	}
-}
-
-// filterUserPublicFields creates a filtered copy of a www User that only
-// contains public information.
-func filterUserPublicFields(user www.User) www.User {
-	return www.User{
-		ID:         user.ID,
-		Admin:      user.Admin,
-		Username:   user.Username,
-		Identities: user.Identities,
-	}
-}
-
-// processUserDetails return the requested user's details. Some fields can be
-// omitted or blank depending on the requester's access level.
-func (p *politeiawww) processUserDetails(ud *www.UserDetails, isCurrentUser bool, isAdmin bool) (*www.UserDetailsReply, error) {
-	// Fetch the database user.
-	user, err := p.userByIDStr(ud.UserID)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert the database user into a proper response.
-	var udr www.UserDetailsReply
-	wwwUser := convertWWWUserFromDatabaseUser(user)
-
-	// Filter returned fields in case the user isn't the admin or the current user
-	if !isAdmin && !isCurrentUser {
-		udr.User = filterUserPublicFields(wwwUser)
-	} else {
-		udr.User = wwwUser
-	}
-
-	return &udr, nil
-}
-
-// processEditUser edits a user's preferences.
-func (p *politeiawww) processEditUser(eu *www.EditUser, user *user.User) (*www.EditUserReply, error) {
-	if eu.EmailNotifications != nil {
-		user.EmailNotifications = *eu.EmailNotifications
-	}
-
-	// Update the user in the database.
-	err := p.db.UserUpdate(*user)
-	if err != nil {
-		return nil, err
-	}
-
-	return &www.EditUserReply{}, nil
-}
-
-// createLoginReply creates a login reply.
-func (p *politeiawww) createLoginReply(u *user.User, lastLoginTime int64) (*www.LoginReply, error) {
-	reply := www.LoginReply{
-		IsAdmin:            u.Admin,
-		UserID:             u.ID.String(),
-		Email:              u.Email,
-		Username:           u.Username,
-		PublicKey:          u.PublicKey(),
-		PaywallAddress:     u.NewUserPaywallAddress,
-		PaywallAmount:      u.NewUserPaywallAmount,
-		PaywallTxNotBefore: u.NewUserPaywallTxNotBefore,
-		PaywallTxID:        u.NewUserPaywallTx,
-		ProposalCredits:    uint64(len(u.UnspentProposalCredits)),
-		LastLoginTime:      lastLoginTime,
-		TOTPVerified:       u.TOTPVerified,
-	}
-
-	if !p.userHasPaid(*u) {
-		err := p.GenerateNewUserPaywall(u)
-		if err != nil {
-			return nil, err
-		}
-
-		reply.PaywallAddress = u.NewUserPaywallAddress
-		reply.PaywallAmount = u.NewUserPaywallAmount
-		reply.PaywallTxNotBefore = u.NewUserPaywallTxNotBefore
-	}
-
-	return &reply, nil
-}
-
-// processVerifyNewUser verifies the token generated for a recently created
-// user.  It ensures that the token matches with the input and that the token
-// hasn't expired.  On success it returns database user record.
-func (p *politeiawww) processVerifyNewUser(usr www.VerifyNewUser) (*user.User, error) {
-	// Check that the user already exists.
-	u, err := p.userByEmail(usr.Email)
-	if err != nil {
-		if errors.Is(err, user.ErrUserNotFound) {
-			log.Debugf("VerifyNewUser failure for %v: user not found",
-				usr.Email)
-			return nil, www.UserError{
-				ErrorCode: www.ErrorStatusVerificationTokenInvalid,
-			}
-		}
-		return nil, err
-	}
-
-	// Decode the verification token.
-	token, err := hex.DecodeString(usr.VerificationToken)
-	if err != nil {
-		log.Debugf("VerifyNewUser failure for %v: verification token could "+
-			"not be decoded: %v", u.Email, err)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenInvalid,
-		}
-	}
-
-	// Check that the verification token matches.
-	if !bytes.Equal(token, u.NewUserVerificationToken) {
-		log.Debugf("VerifyNewUser: wrong token for user %v "+
-			"got %v, want %v", u.Email, hex.EncodeToString(token),
-			hex.EncodeToString(u.NewUserVerificationToken))
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenInvalid,
-		}
-	}
-
-	// Check that the token hasn't expired.
-	if time.Now().Unix() > u.NewUserVerificationExpiry {
-		log.Debugf("VerifyNewUser failure for %v: verification token expired",
-			u.Email)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenExpired,
-		}
-	}
-
-	// Check signature
-	sig, err := util.ConvertSignature(usr.Signature)
-	if err != nil {
-		log.Debugf("VerifyNewUser failure for %v: signature could not be "+
-			"decoded: %v", u.Email, err)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidSignature,
-		}
-	}
-	id := u.InactiveIdentity()
-	if id == nil {
-		log.Debugf("VerifyNewUser failure for %v: no public key",
-			u.Email)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusNoPublicKey,
-		}
-	}
-	pi, err := identity.PublicIdentityFromBytes(id.Key[:])
-	if err != nil {
-		return nil, err
-	}
-	if !pi.VerifyMessage([]byte(usr.VerificationToken), sig) {
-		log.Debugf("VerifyNewUser failure for %v: signature doesn't match "+
-			"(pubkey: %v)", u.Email, pi.String())
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidSignature,
-		}
-	}
-
-	// Clear out the verification token fields
-	// and activate the identity for the user.
-	u.NewUserVerificationToken = nil
-	u.NewUserVerificationExpiry = 0
-	u.ResendNewUserVerificationExpiry = 0
-	err = u.ActivateIdentity(id.Key[:])
-	if err != nil {
-		return nil, err
-	}
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return nil, err
-	}
-
-	p.addUserToPaywallPoolLock(u, paywallTypeUser)
-
-	return u, nil
-}
-
-// processResendVerification resends a new user verification email if the
-// user exists and his verification token is expired.
-func (p *politeiawww) processResendVerification(rv *www.ResendVerification) (*www.ResendVerificationReply, error) {
-	rvr := www.ResendVerificationReply{}
-
-	// Get user from db.
-	u, err := p.userByEmail(rv.Email)
-	if err != nil {
-		if errors.Is(err, user.ErrUserNotFound) {
-			log.Debugf("ResendVerification failure for %v: user not found",
-				rv.Email)
-			return nil, www.UserError{
-				ErrorCode: www.ErrorStatusUserNotFound,
-			}
-		}
-		return nil, err
-	}
-
-	// Don't do anything if the user is already verified or the token hasn't
-	// expired yet.
-	if u.NewUserVerificationToken == nil {
-		log.Debugf("ResendVerification failure for %v: user already verified",
-			rv.Email)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusEmailAlreadyVerified,
-		}
-	}
-
-	if u.ResendNewUserVerificationExpiry > time.Now().Unix() {
-		log.Debugf("ResendVerification failure for %v: verification token "+
-			"not expired yet", rv.Email)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenUnexpired,
-		}
-	}
-
-	// Ensure we got a proper pubkey.
-	err = validatePubKey(rv.PublicKey)
-	if err != nil {
-		return nil, err
-	}
-
-	// Ensure the pubkey is unique
-	_, err = p.db.UserGetByPubKey(rv.PublicKey)
-	switch err {
-	case user.ErrUserNotFound:
-	// Pubkey is unique; continue
-	case nil:
-		// Pubkey is not unique. The user is allowed to use the
-		// same pubkey that they originally signed up with, so
-		// only throw an error if the pubkey is not the user's
-		// inactive identity pubkey.
-		if u.InactiveIdentity().String() != rv.PublicKey {
-			return nil, www.UserError{
-				ErrorCode: www.ErrorStatusDuplicatePublicKey,
-			}
-		}
-	default:
-		// All other errors
-		return nil, err
-	}
-
-	// Set a new verificaton token and identity.
-	token, expiry, err := newVerificationTokenAndExpiry()
-	if err != nil {
-		return nil, err
-	}
-	u.NewUserVerificationToken = token
-	u.NewUserVerificationExpiry = expiry
-	u.ResendNewUserVerificationExpiry = expiry
-	id, err := user.NewIdentity(rv.PublicKey)
-	if err != nil {
-		return nil, err
-	}
-	err = u.AddIdentity(*id)
-	if err != nil {
-		return nil, err
-	}
-
-	// Try to email the verification link first; if it fails, then
-	// the user won't be updated.
-	//
-	// This is conditional on the email server being setup.
-	err = p.emailUserEmailVerify(u.Email,
-		hex.EncodeToString(token), u.Username)
-	if err != nil {
-		log.Errorf("processResendVerification: email verification "+
-			"token failed for '%v': %v", u.Email, err)
-		return nil, err
-	}
-
-	// Update the user in the db.
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return nil, err
-	}
-
-	// Only set the token if email verification is disabled.
-	if !p.mail.IsEnabled() {
-		rvr.VerificationToken = hex.EncodeToString(token)
-	}
-	return &rvr, nil
-}
-
-// processUpdateUserKey sets a verification token and expiry to allow the user
-// to update his key pair; the token must be verified before it expires. If the
-// token is already set and is expired, it generates a new one.
-func (p *politeiawww) processUpdateUserKey(usr *user.User, uuk www.UpdateUserKey) (*www.UpdateUserKeyReply, error) {
-	// Ensure we got a proper pubkey that is unique.
-	err := validatePubKey(uuk.PublicKey)
-	if err != nil {
-		return nil, err
-	}
-	_, err = p.db.UserGetByPubKey(uuk.PublicKey)
-	switch err {
-	case user.ErrUserNotFound:
-		// Pubkey is unique; continue
-	case nil:
-		// Pubkey is not unique
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusDuplicatePublicKey,
-		}
-	default:
-		// All other errors
-		return nil, err
-	}
-
-	// Check if the verification token hasn't expired yet.
-	if usr.UpdateKeyVerificationToken != nil {
-		if usr.UpdateKeyVerificationExpiry > time.Now().Unix() {
-			return nil, www.UserError{
-				ErrorCode: www.ErrorStatusVerificationTokenUnexpired,
-				ErrorContext: []string{
-					strconv.FormatInt(usr.UpdateKeyVerificationExpiry, 10),
-				},
-			}
-		}
-	}
-
-	// Generate a new verification token and expiry.
-	tokenb, expiry, err := newVerificationTokenAndExpiry()
-	if err != nil {
-		return nil, err
-	}
-	usr.UpdateKeyVerificationToken = tokenb
-	usr.UpdateKeyVerificationExpiry = expiry
-
-	// Add inactive identity to the user
-	id, err := user.NewIdentity(uuk.PublicKey)
-	if err != nil {
-		return nil, err
-	}
-	err = usr.AddIdentity(*id)
-	if err != nil {
-		return nil, err
-	}
-
-	// Email the user a verification link. The database does not get
-	// updated if this fails.
-	//
-	// This is conditional on the email server being setup.
-	token := hex.EncodeToString(tokenb)
-	recipient := map[uuid.UUID]string{
-		usr.ID: usr.Email,
-	}
-	err = p.emailUserKeyUpdate(usr.Username, uuk.PublicKey, token, recipient)
-	if err != nil {
-		return nil, err
-	}
-
-	// Save user changes to the database
-	err = p.db.UserUpdate(*usr)
-	if err != nil {
-		return nil, err
-	}
-
-	// Only set the token if email verification is disabled.
-	var t string
-	if !p.mail.IsEnabled() {
-		t = token
-	}
-	return &www.UpdateUserKeyReply{
-		VerificationToken: t,
-	}, nil
-}
-
-// processVerifyUpdateUserKey verifies the token generated for the recently
-// generated key pair. It ensures that the token matches with the input and
-// that the token hasn't expired.
-func (p *politeiawww) processVerifyUpdateUserKey(u *user.User, vu www.VerifyUpdateUserKey) (*user.User, error) {
-	// Decode the verification token.
-	token, err := hex.DecodeString(vu.VerificationToken)
-	if err != nil {
-		log.Debugf("VerifyUpdateUserKey failure for %v: verification "+
-			"token could not be decoded: %v", u.Email, err)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenInvalid,
-		}
-	}
-
-	// Check that the verification token matches.
-	if !bytes.Equal(token, u.UpdateKeyVerificationToken) {
-		log.Debugf("VerifyUpdateUserKey failure for %v: verification "+
-			"token doesn't match, expected %v", u.Email,
-			u.UpdateKeyVerificationToken, token)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenInvalid,
-		}
-	}
-
-	// Check that the token hasn't expired.
-	if u.UpdateKeyVerificationExpiry < time.Now().Unix() {
-		log.Debugf("VerifyUpdateUserKey failure for %v: verification "+
-			"token not expired yet", u.Email)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenExpired,
-		}
-	}
-
-	// Check signature
-	sig, err := util.ConvertSignature(vu.Signature)
-	if err != nil {
-		log.Debugf("VerifyUpdateUserKey failure for %v: signature "+
-			"could not be decoded: %v", u.Email, err)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidSignature,
-		}
-	}
-
-	id := u.InactiveIdentity()
-	if id == nil {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusNoPublicKey,
-		}
-	}
-	pi, err := identity.PublicIdentityFromBytes(id.Key[:])
-	if err != nil {
-		return nil, err
-	}
-
-	if !pi.VerifyMessage([]byte(vu.VerificationToken), sig) {
-		log.Debugf("VerifyUpdateUserKey failure for %v: signature did"+
-			" not match (pubkey: %v)", u.Email, pi.String())
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidSignature,
-		}
-	}
-
-	// Clear out the verification token fields in the db and activate
-	// the key and deactivate the one it's replacing.
-	u.UpdateKeyVerificationToken = nil
-	u.UpdateKeyVerificationExpiry = 0
-	err = u.ActivateIdentity(id.Key[:])
-	if err != nil {
-		return nil, err
-	}
-
-	return u, p.db.UserUpdate(*u)
-}
-
-func (p *politeiawww) login(l www.Login) loginResult {
-	// Get user record
-	u, err := p.userByEmail(l.Email)
-	if err != nil {
-		if errors.Is(err, user.ErrUserNotFound) {
-			log.Debugf("login: user not found for email '%v'",
-				l.Email)
-			err = www.UserError{
-				ErrorCode: www.ErrorStatusInvalidLogin,
-			}
-		}
-		return loginResult{
-			reply: nil,
-			err:   err,
-		}
-	}
-
-	// First check if TOTP is enabled and verified.
-	if u.TOTPVerified {
-		err := p.totpCheck(l.Code, u)
-		if err != nil {
-			return loginResult{
-				reply: nil,
-				err:   err,
-			}
-		}
-	}
-
-	// Verify password
-	err = bcrypt.CompareHashAndPassword(u.HashedPassword,
-		[]byte(l.Password))
-	if err != nil {
-		// Wrong password. Update user record with failed attempt.
-		log.Debugf("login: wrong password")
-		if !userIsLocked(u.FailedLoginAttempts) {
-			u.FailedLoginAttempts++
-			u.TOTPLastFailedCodeTime = make([]int64, 0, 2)
-			err := p.db.UserUpdate(*u)
-			if err != nil {
-				return loginResult{
-					reply: nil,
-					err:   err,
-				}
-			}
-			// If the failed attempt puts the user over the limit,
-			// send them an email informing them their account is
-			// now locked.
-			if userIsLocked(u.FailedLoginAttempts) {
-				recipient := map[uuid.UUID]string{
-					u.ID: u.Email,
-				}
-				err := p.emailUserAccountLocked(u.Username, recipient)
-				if err != nil {
-					return loginResult{
-						reply: nil,
-						err:   err,
-					}
-				}
-			}
-		}
-		return loginResult{
-			reply: nil,
-			err: www.UserError{
-				ErrorCode: www.ErrorStatusInvalidLogin,
-			},
-		}
-	}
-
-	// Verify user account is in good standing
-	if u.NewUserVerificationToken != nil {
-		return loginResult{
-			reply: nil,
-			err: www.UserError{
-				ErrorCode: www.ErrorStatusEmailNotVerified,
-			},
-		}
-	}
-	if u.Deactivated {
-		return loginResult{
-			reply: nil,
-			err: www.UserError{
-				ErrorCode: www.ErrorStatusUserDeactivated,
-			},
-		}
-	}
-	if userIsLocked(u.FailedLoginAttempts) {
-		return loginResult{
-			reply: nil,
-			err: www.UserError{
-				ErrorCode: www.ErrorStatusUserLocked,
-			},
-		}
-	}
-
-	// Update user record with successful login
-	lastLoginTime := u.LastLoginTime
-	u.FailedLoginAttempts = 0
-	u.LastLoginTime = time.Now().Unix()
-	u.TOTPLastFailedCodeTime = make([]int64, 0, 2)
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return loginResult{
-			reply: nil,
-			err:   err,
-		}
-	}
-
-	reply, err := p.createLoginReply(u, lastLoginTime)
-	return loginResult{
-		reply: reply,
-		err:   err,
-	}
-}
-
-// processLogin logs the provided user into politeia. This is done using go
-// routines in order to prevent timing attacks.
-func (p *politeiawww) processLogin(l www.Login) (*www.LoginReply, error) {
-	log.Tracef("processLogin: %v", l.Email)
-
-	var (
-		wg sync.WaitGroup
-		ch = make(chan loginResult)
-	)
-
-	// Wait for both go routines to finish before returning the
-	// reply. This is done to prevent an attacker from being able
-	// to execute a timing attack to determine if the provided
-	// email address is the user's valid email address.
-	wg.Add(2)
-	go func() {
-		defer wg.Done()
-		ch <- p.login(l)
-	}()
-	go func() {
-		defer wg.Done()
-		time.Sleep(loginMinWaitTime)
-	}()
-	lr := <-ch
-	wg.Wait()
-
-	return lr.reply, lr.err
-}
-
-/*
-XXX this login implementation is being commented out until we switch
-the login credentials back to using username instead of email.
-https://github.com/decred/politeia/issues/860#issuecomment-520871500
-
-// processLogin checks that the provided user credentials are valid and updates
-// the login fields for the user.
-func (p *politeiawww) processLogin(l www.Login) (*www.LoginReply, error) {
-	log.Tracef("processLogin: %v", l.Username)
-
-	// Lookup user
-	u, err := p.db.UserGetByUsername(l.Username)
-	if err != nil {
-		if errors.Is(err, user.ErrUserNotFound) {
-			err = www.UserError{
-				ErrorCode: www.ErrorStatusUserNotFound,
-			}
-		}
-		return nil, err
-	}
-
-	// Ensure the account isn't locked
-	if userIsLocked(u.FailedLoginAttempts) {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusUserLocked,
-		}
-	}
-
-	// Check the password
-	err = bcrypt.CompareHashAndPassword(u.HashedPassword,
-		[]byte(l.Password))
-	if err != nil {
-		// Password is wrong. Increment failed login attempts.
-		u.FailedLoginAttempts++
-		err := p.db.UserUpdate(*u)
-		if err != nil {
-			return nil, err
-		}
-
-		// If the user account has reached the limit for failed
-		// login attempts, send the user an email to notify them
-		// that their account is locked.
-		if userIsLocked(u.FailedLoginAttempts) {
-			err := p.emailUserAccountLocked(u.Email)
-			if err != nil {
-				log.Errorf("processLogin: emailUserAccountLocked '%v': %v",
-					u.Email, err)
-			}
-		}
-
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidPassword,
-		}
-	}
-
-	// Ensure user has been verified and has not been deactivated.
-	if u.NewUserVerificationToken != nil {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusEmailNotVerified,
-		}
-	}
-	if u.Deactivated {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusUserDeactivated,
-		}
-	}
-
-	// Update user login fields
-	lastLoginTime := u.LastLoginTime
-	u.FailedLoginAttempts = 0
-	u.LastLoginTime = time.Now().Unix()
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return nil, err
-	}
-
-	return p.createLoginReply(u, lastLoginTime)
-}
-*/
-
-// processChangeUsername checks that the password matches the one
-// in the database, then checks that the username is valid and not
-// already taken, then changes the user record in the database to
-// the new username.
-func (p *politeiawww) processChangeUsername(email string, cu www.ChangeUsername) (*www.ChangeUsernameReply, error) {
-	var reply www.ChangeUsernameReply
-
-	// Get user from db.
-	u, err := p.userByEmail(email)
-	if err != nil {
-		return nil, err
-	}
-
-	// Check the user's password.
-	err = bcrypt.CompareHashAndPassword(u.HashedPassword,
-		[]byte(cu.Password))
-	if err != nil {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidPassword,
-		}
-	}
-
-	// Format and validate the new username.
-	newUsername := formatUsername(cu.NewUsername)
-	err = validateUsername(newUsername)
-	if err != nil {
-		return nil, err
-	}
-
-	// Check for duplicate username
-	_, err = p.db.UserGetByUsername(newUsername)
-	switch err {
-	case nil:
-		// Duplicate
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusDuplicateUsername,
-		}
-	case user.ErrUserNotFound:
-		// Doesn't exist, update username.
-	default:
-		// All other errors
-		return nil, err
-	}
-
-	// Add the updated user information to the db.
-	u.Username = newUsername
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return nil, err
-	}
-
-	return &reply, nil
-}
-
-// processChangePassword checks that the current password matches the one
-// in the database, then changes it to the new password.
-func (p *politeiawww) processChangePassword(email string, cp www.ChangePassword) (*www.ChangePasswordReply, error) {
-	var reply www.ChangePasswordReply
-
-	// Get user from db.
-	u, err := p.userByEmail(email)
-	if err != nil {
-		return nil, err
-	}
-
-	// Check the user's password.
-	err = bcrypt.CompareHashAndPassword(u.HashedPassword,
-		[]byte(cp.CurrentPassword))
-	if err != nil {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidPassword,
-		}
-	}
-
-	// Validate the new password.
-	err = validatePassword(cp.NewPassword)
-	if err != nil {
-		return nil, err
-	}
-
-	// Hash the user's password.
-	hashedPassword, err := p.hashPassword(cp.NewPassword)
-	if err != nil {
-		return nil, err
-	}
-
-	// Add the updated user information to the db.
-	u.HashedPassword = hashedPassword
-
-	// We will also reset any possibly issued verification token to avoid
-	// a small chance of one having been issued by a potential attacker.
-	// Any update to the password by a logged in user, should be seen as
-	// an authorized request and therefore override any potential request.
-	u.ResetPasswordVerificationToken = nil
-	u.ResetPasswordVerificationExpiry = 0
-
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return nil, err
-	}
-
-	recipient := map[uuid.UUID]string{
-		u.ID: u.Email,
-	}
-	err = p.emailUserPasswordChanged(u.Username, recipient)
-	if err != nil {
-		return nil, err
-	}
-
-	return &reply, nil
-}
-
-func (p *politeiawww) resetPassword(rp www.ResetPassword) resetPasswordResult {
-	// Lookup user
-	u, err := p.db.UserGetByUsername(rp.Username)
-	if err != nil {
-		if errors.Is(err, user.ErrUserNotFound) {
-			err = www.UserError{
-				ErrorCode: www.ErrorStatusUserNotFound,
-			}
-		}
-		return resetPasswordResult{
-			err: err,
-		}
-	}
-
-	// Ensure the provided email address matches the user record
-	// email address. If the addresses does't match, return so
-	// that the verification token doesn't get sent.
-	if rp.Email != u.Email {
-		log.Debugf("resetPassword: wrong email: %v %v",
-			rp.Email, u.Email)
-		return resetPasswordResult{}
-	}
-
-	// If the user already has a verification token that has not
-	// yet expired, do nothing.
-	t := time.Now().Unix()
-	if t < u.ResetPasswordVerificationExpiry {
-		log.Debugf("resetPassword: unexpired verification token: %v %v",
-			t, u.ResetPasswordVerificationExpiry)
-		return resetPasswordResult{}
-	}
-
-	// The verification token is not present or is present but has expired.
-
-	// Generate a new verification token and expiry.
-	tokenb, expiry, err := newVerificationTokenAndExpiry()
-	if err != nil {
-		return resetPasswordResult{
-			err: err,
-		}
-	}
-
-	// Try to email the verification link first. If it fails, the
-	// user record won't be updated in the database.
-	recipient := map[uuid.UUID]string{
-		u.ID: u.Email,
-	}
-	err = p.emailUserPasswordReset(rp.Username, hex.EncodeToString(tokenb),
-		recipient)
-	if err != nil {
-		return resetPasswordResult{
-			err: err,
-		}
-	}
-
-	// Update the user record
-	u.ResetPasswordVerificationToken = tokenb
-	u.ResetPasswordVerificationExpiry = expiry
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return resetPasswordResult{
-			err: err,
-		}
-	}
-
-	// Only include the verification token in the reply if the
-	// email server has been disabled.
-	var reply www.ResetPasswordReply
-	if !p.mail.IsEnabled() {
-		reply.VerificationToken = hex.EncodeToString(tokenb)
-	}
-
-	return resetPasswordResult{
-		reply: reply,
-	}
-}
-
-// processResetPassword is used to perform a password change when the user is
-// not logged in. The provided email address must match the email address
-// or the user record that corresponds to the provided username.
-func (p *politeiawww) processResetPassword(rp www.ResetPassword) (*www.ResetPasswordReply, error) {
-	log.Tracef("processResetPassword: %v", rp.Username)
-	var (
-		wg sync.WaitGroup
-		ch = make(chan resetPasswordResult)
-	)
-
-	// Wait for both go routines to finish before returning the
-	// reply. This is done to prevent an attacker from being able
-	// to execute a timing attack to determine if the provided
-	// email address is the user's valid email address.
-	wg.Add(2)
-	go func() {
-		defer wg.Done()
-		ch <- p.resetPassword(rp)
-	}()
-	go func() {
-		defer wg.Done()
-		time.Sleep(resetPasswordMinWaitTime)
-	}()
-	rpr := <-ch
-	wg.Wait()
-
-	return &rpr.reply, rpr.err
-}
-
-// processVerifyResetPassword verifies the token that was sent to the user
-// during the reset password command. If everything checks out, the user's
-// password is updated with the provided new password and the user's account
-// is unlocked if it had previously been locked.
-func (p *politeiawww) processVerifyResetPassword(vrp www.VerifyResetPassword) (*www.VerifyResetPasswordReply, error) {
-	log.Tracef("processVerifyResetPassword: %v %v",
-		vrp.Username, vrp.VerificationToken)
-
-	// Lookup user
-	u, err := p.db.UserGetByUsername(vrp.Username)
-	if err != nil {
-		if errors.Is(err, user.ErrUserNotFound) {
-			err = www.UserError{
-				ErrorCode: www.ErrorStatusUserNotFound,
-			}
-		}
-		return nil, err
-	}
-
-	// Validate verification token
-	token, err := hex.DecodeString(vrp.VerificationToken)
-	if err != nil {
-		log.Debugf("processVerifyResetPassword: decode hex '%v': %v",
-			vrp.VerificationToken, err)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenInvalid,
-		}
-	}
-	if !bytes.Equal(token, u.ResetPasswordVerificationToken) {
-		log.Debugf("processVerifyResetPassword: wrong token: %v %v",
-			hex.EncodeToString(token),
-			hex.EncodeToString(u.ResetPasswordVerificationToken))
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenInvalid,
-		}
-	}
-	if u.ResetPasswordVerificationExpiry < time.Now().Unix() {
-		log.Debugf("processVerifyResetPassword: token expired: %v %v",
-			u.ResetPasswordVerificationExpiry, time.Now().Unix())
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVerificationTokenExpired,
-		}
-	}
-
-	// The client should be hashing the password before sending
-	// it to politeiawww. This validation is only relevant if the
-	// client failed to hash the password or does not include a
-	// password in the request.
-	err = validatePassword(vrp.NewPassword)
-	if err != nil {
-		return nil, err
-	}
-
-	// Hash the new password
-	hashedPassword, err := p.hashPassword(vrp.NewPassword)
-	if err != nil {
-		return nil, err
-	}
-
-	// Update the user
-	u.ResetPasswordVerificationToken = nil
-	u.ResetPasswordVerificationExpiry = 0
-	u.HashedPassword = hashedPassword
-	u.FailedLoginAttempts = 0
-
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return nil, err
-	}
-
-	return &www.VerifyResetPasswordReply{}, nil
-}
-
-func convertProposalCreditFromUserDB(credit user.ProposalCredit) www.ProposalCredit {
-	return www.ProposalCredit{
-		PaywallID:     credit.PaywallID,
-		Price:         credit.Price,
-		DatePurchased: credit.DatePurchased,
-		TxID:          credit.TxID,
-	}
-}
 
 // _logAdminAction logs a string to the admin log file.
 //
 // This function must be called WITH the mutex held.
-func (p *politeiawww) _logAdminAction(adminUser *user.User, content string) error {
+func (p *LegacyPoliteiawww) _logAdminAction(adminUser *user.User, content string) error {
 	if p.test {
 		return nil
 	}
@@ -1067,7 +36,7 @@ func (p *politeiawww) _logAdminAction(adminUser *user.User, content string) erro
 // logAdminAction logs a string to the admin log file.
 //
 // This function must be called WITHOUT the mutex held.
-func (p *politeiawww) logAdminAction(adminUser *user.User, content string) error {
+func (p *LegacyPoliteiawww) logAdminAction(adminUser *user.User, content string) error {
 	p.Lock()
 	defer p.Unlock()
 
@@ -1077,7 +46,7 @@ func (p *politeiawww) logAdminAction(adminUser *user.User, content string) error
 // logAdminUserAction logs an admin action on a specific user.
 //
 // This function must be called WITHOUT the mutex held.
-func (p *politeiawww) logAdminUserAction(adminUser, user *user.User, action www.UserManageActionT, reasonForAction string) error {
+func (p *LegacyPoliteiawww) logAdminUserAction(adminUser, user *user.User, action www.UserManageActionT, reasonForAction string) error {
 	return p.logAdminAction(adminUser, fmt.Sprintf("%v,%v,%v,%v",
 		www.UserManageAction[action], user.ID, user.Username, reasonForAction))
 }
@@ -1085,537 +54,6 @@ func (p *politeiawww) logAdminUserAction(adminUser, user *user.User, action www.
 // logAdminProposalAction logs an admin action on a proposal.
 //
 // This function must be called WITHOUT the mutex held.
-func (p *politeiawww) logAdminProposalAction(adminUser *user.User, token, action, reason string) error {
+func (p *LegacyPoliteiawww) logAdminProposalAction(adminUser *user.User, token, action, reason string) error {
 	return p.logAdminAction(adminUser, fmt.Sprintf("%v,%v,%v", action, token, reason))
-}
-
-// processManageUser processes the admin ManageUser command.
-func (p *politeiawww) processManageUser(mu *www.ManageUser, adminUser *user.User) (*www.ManageUserReply, error) {
-	// Fetch the database user.
-	user, err := p.userByIDStr(mu.UserID)
-	if err != nil {
-		return nil, err
-	}
-
-	// Validate that the action is valid.
-	if mu.Action == www.UserManageInvalid {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidUserManageAction,
-		}
-	}
-
-	// Validate that the reason is supplied.
-	mu.Reason = strings.TrimSpace(mu.Reason)
-	if len(mu.Reason) == 0 {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidInput,
-		}
-	}
-
-	// -168 hours is 7 days in the past
-	expiredTime := time.Now().Add(-168 * time.Hour).Unix()
-
-	switch mu.Action {
-	case www.UserManageExpireNewUserVerification:
-		user.NewUserVerificationExpiry = expiredTime
-		user.ResendNewUserVerificationExpiry = expiredTime
-	case www.UserManageExpireUpdateKeyVerification:
-		user.UpdateKeyVerificationExpiry = expiredTime
-	case www.UserManageExpireResetPasswordVerification:
-		user.ResetPasswordVerificationExpiry = expiredTime
-	case www.UserManageClearUserPaywall:
-		p.removeUsersFromPool([]uuid.UUID{user.ID}, paywallTypeUser)
-		user.NewUserPaywallAmount = 0
-		user.NewUserPaywallTx = "cleared_by_admin"
-		user.NewUserPaywallPollExpiry = 0
-	case www.UserManageUnlock:
-		user.FailedLoginAttempts = 0
-	case www.UserManageDeactivate:
-		user.Deactivated = true
-	case www.UserManageReactivate:
-		user.Deactivated = false
-	default:
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidUserManageAction,
-		}
-	}
-
-	// Update the user in the database.
-	err = p.db.UserUpdate(*user)
-	if err != nil {
-		return nil, err
-	}
-
-	return &www.ManageUserReply{}, nil
-}
-
-// processUsers returns a list of users given a set of filters. Admins can
-// search by pubkey, username or email. Username and email searches will
-// return partial matches. Pubkey searches must be an exact match. Non admins
-// can search by pubkey or username. Non admin searches will only return exact
-// matches.
-func (p *politeiawww) processUsers(users *www.Users, isAdmin bool) (*www.UsersReply, error) {
-	log.Tracef("processUsers: %v", isAdmin)
-
-	emailQuery := strings.ToLower(users.Email)
-	usernameQuery := formatUsername(users.Username)
-	pubkeyQuery := users.PublicKey
-
-	var u *user.User
-	var totalUsers uint64
-	var totalMatches uint64
-	var pubkeyMatchID string
-	matchedUsers := make([]www.AbridgedUser, 0, www.UserListPageSize)
-
-	if pubkeyQuery != "" {
-		// Search by pubkey. Only exact matches are returned.
-		// Validate pubkey
-		err := validatePubKey(pubkeyQuery)
-		if err != nil {
-			return nil, err
-		}
-
-		u, err = p.db.UserGetByPubKey(pubkeyQuery)
-		if err != nil {
-			if errors.Is(err, user.ErrUserNotFound) {
-				// Pubkey searches require an exact match. If no
-				// match was found, we can go ahead and return.
-				return &www.UsersReply{}, nil
-			}
-			return nil, err
-		}
-
-		pubkeyMatchID = u.ID.String()
-	}
-
-	switch {
-	case isAdmin:
-		// Admins can search by username and/or email with
-		// partial matches being returned.
-		err := p.db.AllUsers(func(user *user.User) {
-			totalUsers++
-			userMatches := true
-
-			// If both emailQuery and usernameQuery are non-empty, the
-			// user must match both to be included in the results.
-			if emailQuery != "" {
-				if !strings.Contains(strings.ToLower(user.Email),
-					emailQuery) {
-					userMatches = false
-				}
-			}
-
-			if usernameQuery != "" && userMatches {
-				if !strings.Contains(strings.ToLower(user.Username),
-					usernameQuery) {
-					userMatches = false
-				}
-			}
-
-			if pubkeyQuery != "" && userMatches {
-				if user.ID.String() != pubkeyMatchID {
-					userMatches = false
-				}
-			}
-
-			if userMatches {
-				totalMatches++
-				if totalMatches < www.UserListPageSize {
-					matchedUsers = append(matchedUsers, www.AbridgedUser{
-						ID:       user.ID.String(),
-						Email:    user.Email,
-						Username: user.Username,
-					})
-				}
-			}
-		})
-		if err != nil {
-			return nil, err
-		}
-
-		// Sort results alphabetically.
-		sort.Slice(matchedUsers, func(i, j int) bool {
-			return matchedUsers[i].Username < matchedUsers[j].Username
-		})
-
-	default:
-		// Non-admins can search by username and the search
-		// must be an exact match.
-		if usernameQuery != "" {
-			// Validate username
-			err := validateUsername(usernameQuery)
-			if err != nil {
-				return nil, err
-			}
-
-			u, err = p.db.UserGetByUsername(usernameQuery)
-			if err != nil {
-				// ErrUserNotFound is ok. Empty search results
-				// will be returned.
-				if !errors.Is(err, user.ErrUserNotFound) {
-					return nil, err
-				}
-			}
-
-			// If both pubkeyQuery and usernameQuery are non-empty, the
-			// user must match both to be included in the results.
-			if (u != nil) && (pubkeyQuery != "") &&
-				(u.ID.String() != pubkeyMatchID) {
-				// User doesn't match both
-				u = nil
-			}
-		}
-
-		if u != nil {
-			totalMatches++
-			matchedUsers = append(matchedUsers, www.AbridgedUser{
-				ID:       u.ID.String(),
-				Username: u.Username})
-		}
-	}
-
-	return &www.UsersReply{
-		TotalUsers:   totalUsers,
-		TotalMatches: totalMatches,
-		Users:        matchedUsers,
-	}, nil
-}
-
-// processUserRegistrationPayment verifies that the provided transaction
-// meets the minimum requirements to mark the user as paid, and then does
-// that in the user database.
-func (p *politeiawww) processUserRegistrationPayment(ctx context.Context, u *user.User) (*www.UserRegistrationPaymentReply, error) {
-	var reply www.UserRegistrationPaymentReply
-	if p.userHasPaid(*u) {
-		reply.HasPaid = true
-		return &reply, nil
-	}
-
-	if paywallHasExpired(u.NewUserPaywallPollExpiry) {
-		err := p.GenerateNewUserPaywall(u)
-		if err != nil {
-			return nil, err
-		}
-		reply.PaywallAddress = u.NewUserPaywallAddress
-		reply.PaywallAmount = u.NewUserPaywallAmount
-		reply.PaywallTxNotBefore = u.NewUserPaywallTxNotBefore
-		return &reply, nil
-	}
-
-	tx, _, err := util.FetchTxWithBlockExplorers(ctx, p.params,
-		u.NewUserPaywallAddress, u.NewUserPaywallAmount,
-		u.NewUserPaywallTxNotBefore, p.cfg.MinConfirmationsRequired,
-		p.dcrdataHostHTTP())
-	if err != nil {
-		return nil, err
-	}
-
-	if tx != "" {
-		reply.HasPaid = true
-
-		err = p.updateUserAsPaid(u, tx)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		// TODO: Add the user to the in-memory pool.
-	}
-
-	return &reply, nil
-}
-
-// processUserProposalPaywall returns a proposal paywall that enables the
-// the user to purchase proposal credits. The user can only have one paywall
-// active at a time.  If no paywall currently exists, a new one is created and
-// the user is added to the paywall pool.
-func (p *politeiawww) processUserProposalPaywall(u *user.User) (*www.UserProposalPaywallReply, error) {
-	log.Tracef("processUserProposalPaywall")
-
-	// Ensure paywall is enabled
-	if !p.paywallIsEnabled() {
-		return &www.UserProposalPaywallReply{}, nil
-	}
-
-	// Proposal paywalls cannot be generated until the user has paid their
-	// user registration fee.
-	if !p.userHasPaid(*u) {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusUserNotPaid,
-		}
-	}
-
-	var pp *user.ProposalPaywall
-	if p.userHasValidProposalPaywall(u) {
-		// Don't create a new paywall if a valid one already exists.
-		pp = p.mostRecentProposalPaywall(u)
-	} else {
-		// Create a new paywall.
-		var err error
-		pp, err = p.generateProposalPaywall(u)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return &www.UserProposalPaywallReply{
-		CreditPrice:        pp.CreditPrice,
-		PaywallAddress:     pp.Address,
-		PaywallTxNotBefore: pp.TxNotBefore,
-	}, nil
-}
-
-// processUserProposalPaywallTx checks if the user has a pending paywall
-// payment and returns the payment details if one is found.
-func (p *politeiawww) processUserProposalPaywallTx(u *user.User) (*www.UserProposalPaywallTxReply, error) {
-	log.Tracef("processUserProposalPaywallTx")
-
-	var (
-		txID          string
-		txAmount      uint64
-		confirmations uint64
-	)
-
-	p.RLock()
-	defer p.RUnlock()
-
-	poolMember, ok := p.userPaywallPool[u.ID]
-	if ok {
-		txID = poolMember.txID
-		txAmount = poolMember.txAmount
-		confirmations = poolMember.txConfirmations
-	}
-
-	return &www.UserProposalPaywallTxReply{
-		TxID:          txID,
-		TxAmount:      txAmount,
-		Confirmations: confirmations,
-	}, nil
-}
-
-// processUserProposalCredits returns a list of the user's unspent proposal
-// credits and a list of the user's spent proposal credits.
-func processUserProposalCredits(u *user.User) (*www.UserProposalCreditsReply, error) {
-	// Convert from database proposal credits to www proposal credits.
-	upc := make([]www.ProposalCredit, len(u.UnspentProposalCredits))
-	for i, credit := range u.UnspentProposalCredits {
-		upc[i] = convertProposalCreditFromUserDB(credit)
-	}
-	spc := make([]www.ProposalCredit, len(u.SpentProposalCredits))
-	for i, credit := range u.SpentProposalCredits {
-		spc[i] = convertProposalCreditFromUserDB(credit)
-	}
-
-	return &www.UserProposalCreditsReply{
-		UnspentCredits: upc,
-		SpentCredits:   spc,
-	}, nil
-}
-
-// processUserPaymentsRescan allows an admin to rescan a user's paywall address
-// to check for any payments that may have been missed by paywall polling.
-func (p *politeiawww) processUserPaymentsRescan(ctx context.Context, upr www.UserPaymentsRescan) (*www.UserPaymentsRescanReply, error) {
-	// Ensure paywall is enabled
-	if !p.paywallIsEnabled() {
-		return &www.UserPaymentsRescanReply{}, nil
-	}
-
-	// Lookup user
-	u, err := p.userByIDStr(upr.UserID)
-	if err != nil {
-		return nil, err
-	}
-
-	// Fetch user payments
-	payments, err := util.FetchTxsForAddressNotBefore(ctx, p.params, u.NewUserPaywallAddress,
-		u.NewUserPaywallTxNotBefore, p.dcrdataHostHTTP())
-	if err != nil {
-		return nil, fmt.Errorf("FetchTxsForAddressNotBefore: %v", err)
-	}
-
-	// Paywalls are in chronological order so sort txs into chronological
-	// order to make them easier to work with
-	sort.SliceStable(payments, func(i, j int) bool {
-		return payments[i].Timestamp < payments[j].Timestamp
-	})
-
-	// Sanity check. Paywalls should already be in chronological order.
-	paywalls := u.ProposalPaywalls
-	sort.SliceStable(paywalls, func(i, j int) bool {
-		return paywalls[i].TxNotBefore < paywalls[j].TxNotBefore
-	})
-
-	// Check for payments that were missed by paywall polling
-	newCredits := make([]user.ProposalCredit, 0, len(payments))
-	for _, payment := range payments {
-		// Check if the payment transaction corresponds to a user
-		// registration payment. A user registration payment may not
-		// exist if the registration paywall was cleared by an admin.
-		if payment.TxID == u.NewUserPaywallTx {
-			continue
-		}
-
-		// Check for credits that correspond to the payment.  If a
-		// credit is found it means that this payment was not missed by
-		// paywall polling and we can continue onto the next payment.
-		var found bool
-		for _, credit := range u.SpentProposalCredits {
-			if credit.TxID == payment.TxID {
-				found = true
-				break
-			}
-		}
-		if found {
-			continue
-		}
-
-		for _, credit := range u.UnspentProposalCredits {
-			if credit.TxID == payment.TxID {
-				found = true
-				break
-			}
-		}
-		if found {
-			continue
-		}
-
-		// Credits were not found for this payment which means that it
-		// was missed by paywall polling. Create new credits using the
-		// paywall details that correspond to the payment timestamp. If
-		// a paywall had not yet been issued, use the current proposal
-		// credit price.
-		var pp user.ProposalPaywall
-		for _, paywall := range paywalls {
-			if payment.Timestamp < paywall.TxNotBefore {
-				continue
-			}
-			if payment.Timestamp > paywall.TxNotBefore {
-				// Corresponding paywall found
-				pp = paywall
-				break
-			}
-		}
-
-		if pp == (user.ProposalPaywall{}) {
-			// Paywall not found. This means the tx occurred before
-			// any paywalls were issued. Use current credit price.
-			pp.CreditPrice = p.cfg.PaywallAmount
-		}
-
-		// Don't add credits if the paywall is in the paywall pool
-		if pp.TxID == "" && !paywallHasExpired(pp.PollExpiry) {
-			continue
-		}
-
-		// Ensure payment has minimum number of confirmations
-		if payment.Confirmations < p.cfg.MinConfirmationsRequired {
-			continue
-		}
-
-		// Create proposal credits
-		numCredits := payment.Amount / pp.CreditPrice
-		c := make([]user.ProposalCredit, numCredits)
-		for i := uint64(0); i < numCredits; i++ {
-			c[i] = user.ProposalCredit{
-				PaywallID:     pp.ID,
-				Price:         pp.CreditPrice,
-				DatePurchased: time.Now().Unix(),
-				TxID:          payment.TxID,
-			}
-		}
-		newCredits = append(newCredits, c...)
-	}
-
-	// Update user record
-	// We relookup the user record here in case the user has spent proposal
-	// credits since the start of this request. Failure to relookup the
-	// user record here could result in adding proposal credits to the
-	// user's account that have already been spent.
-	u, err = p.userByEmail(u.Email)
-	if err != nil {
-		return nil, fmt.Errorf("UserGet %v", err)
-	}
-
-	u.UnspentProposalCredits = append(u.UnspentProposalCredits,
-		newCredits...)
-
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return nil, fmt.Errorf("UserUpdate %v", err)
-	}
-
-	// Convert database credits to www credits
-	newCreditsWWW := make([]www.ProposalCredit, len(newCredits))
-	for i, credit := range newCredits {
-		newCreditsWWW[i] = convertProposalCreditFromUserDB(credit)
-	}
-
-	return &www.UserPaymentsRescanReply{
-		NewCredits: newCreditsWWW,
-	}, nil
-}
-
-func (p *politeiawww) totpCheck(code string, u *user.User) error {
-	// Return error to alert that a code is required.
-	if code == "" {
-		log.Debugf("login: totp code required %v", u.Email)
-		return www.UserError{
-			ErrorCode: www.ErrorStatusRequiresTOTPCode,
-		}
-	}
-
-	// Get the generated totp code. The provided code must match this
-	// generated code.
-	requestTime := time.Now()
-	currentCode, err := p.totpGenerateCode(u.TOTPSecret, requestTime)
-	if err != nil {
-		return fmt.Errorf("totpGenerateCode: %v", err)
-	}
-
-	// Verify the user does not have too many failed attempts for this
-	// epoch.
-	if len(u.TOTPLastFailedCodeTime) >= totpFailedAttempts {
-		// The user has too many failed attempts. We must first verify
-		// that the failed attempts are from this epoch before this is
-		// considered to be an error. If the generated code from the
-		// failed timestamp matches the generated code from the current
-		// timestamp then we know the failures occurred during this epoch.
-		failedTS := u.TOTPLastFailedCodeTime[len(u.TOTPLastFailedCodeTime)-1]
-		oldCode, err := p.totpGenerateCode(u.TOTPSecret, time.Unix(failedTS, 0))
-		if err != nil {
-			return fmt.Errorf("totpGenerateCode: %v", err)
-		}
-		if oldCode == currentCode {
-			// The failures occurred in the same epoch which means the user
-			// has exceeded their max allowed attempts.
-			return www.UserError{
-				ErrorCode: www.ErrorStatusTOTPWaitForNewCode,
-			}
-		}
-
-		// Previous failures are not from this epoch. Clear them out.
-		u.TOTPLastFailedCodeTime = []int64{}
-	}
-
-	// Verify the provided code matches the generated code
-	var replyError error
-	if currentCode == code {
-		// The code matches. Clear out all previous failed attempts
-		// before returning.
-		u.TOTPLastFailedCodeTime = []int64{}
-	} else {
-		// The code doesn't match. Save the failure and return an error.
-		ts := requestTime.Unix()
-		u.TOTPLastFailedCodeTime = append(u.TOTPLastFailedCodeTime, ts)
-		replyError = www.UserError{
-			ErrorCode: www.ErrorStatusTOTPFailedValidation,
-		}
-	}
-
-	// Update the user database
-	err = p.db.UserUpdate(*u)
-	if err != nil {
-		return fmt.Errorf("UserUpdate: %v", err)
-	}
-
-	return replyError
 }
